requirements.txt
asttokens==2.4.1
canopen==2.3.0
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
colorzero==2.0
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
distro==1.8.0
executing==2.0.1
fonttools==4.53.0
gpiozero==2.0
idna==3.3
ipython==8.25.0
jedi==0.19.1
kiwisolver==1.4.5
lgpio==0.2.2.0
matplotlib==3.9.0
matplotlib-inline==0.1.7
msgpack==1.0.8
numpy==2.0.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pigpio==1.78
pillow==10.3.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodomex==3.11.0
Pygments==2.18.0
pyparsing==3.1.2
python-apt==2.6.0
python-can==4.4.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.28.1
RPi.GPIO==0.7.1a4
six==1.16.0
smbus2==0.4.2
spidev==3.5
ssh-import-id==5.10
stack-data==0.6.3
toml==0.10.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==1.26.12
voluptuous==0.14.2
wcwidth==0.2.13
wrapt==1.16.0
